pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nuts-scheduler"
version = "0.4.0"
dependencies = ["python-dateutil==2.9.0.post0", "redis==5.2.1"]
authors = [
  { name="Sam Huffman", email="huffmsa@gmail.com" },
]
description = "A Redis backed job runner"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]

[project.urls]
Homepage = "https://github.com/huffmsa/nuts"
Issues = "https://github.com/huffmsa/nuts/issues"
